--- src/think/db/concern/UpdateCheck.php
@@ -1,95 +1,0 @@
-<?php
-// +----------------------------------------------------------------------
-// | ThinkPHP [ WE CAN DO IT JUST THINK ]
-// +----------------------------------------------------------------------
-// | Copyright (c) 2006~2019 http://thinkphp.cn All rights reserved.
-// +----------------------------------------------------------------------
-// | Licensed ( http://www.apache.org/licenses/LICENSE-2.0 )
-// +----------------------------------------------------------------------
-// | Author: liu21st <liu21st@gmail.com>
-// +----------------------------------------------------------------------
-declare (strict_types = 1);
-
-namespace think\db\concern;
-
-use think\db\Exception;
-
-trait UpdateCheck
-{
-    /**
-     * 分析数据是否存在更新条件
-     * @access public
-     * @param array $data 数据
-     * @return bool
-     * @throws Exception
-     */
-    public function parseUpdateData(&$data): bool
-    {
-        $pk       = $this->getPk();
-        $isUpdate = false;
-        // 如果存在主键数据 则自动作为更新条件
-        if (is_string($pk) && isset($data[$pk])) {
-            $this->where($pk, '=', $data[$pk]);
-            $this->options['key'] = $data[$pk];
-            unset($data[$pk]);
-            $isUpdate = true;
-        } elseif (is_array($pk)) {
-            foreach ($pk as $field) {
-                if (isset($data[$field])) {
-                    $this->where($field, '=', $data[$field]);
-                    $isUpdate = true;
-                } else {
-                    // 如果缺少复合主键数据则不执行
-                    throw new Exception('miss complex primary data');
-                }
-                unset($data[$field]);
-            }
-        }
-
-        return $isUpdate;
-    }
-
-    /**
-     * 把主键值转换为查询条件 支持复合主键
-     * @access public
-     * @param array|string $data 主键数据
-     * @return void
-     * @throws Exception
-     */
-    public function parsePkWhere($data): void
-    {
-        $pk = $this->getPk();
-
-        if (is_string($pk)) {
-            // 获取数据表
-            if (empty($this->options['table'])) {
-                $this->options['table'] = $this->getTable();
-            }
-
-            $table = is_array($this->options['table']) ? key($this->options['table']) : $this->options['table'];
-
-            if (!empty($this->options['alias'][$table])) {
-                $alias = $this->options['alias'][$table];
-            }
-
-            $key = isset($alias) ? $alias . '.' . $pk : $pk;
-            // 根据主键查询
-            if (is_array($data)) {
-                $this->where($key, 'in', $data);
-            } else {
-                $this->where($key, '=', $data);
-                $this->options['key'] = $data;
-            }
-        }
-    }
-
-    /**
-     * 获取模型的更新条件
-     * @access protected
-     * @param array $options 查询参数
-     */
-    protected function getModelUpdateCondition(array $options)
-    {
-        return $options['where']['AND'] ?? null;
-    }
-}
